node: correct transferables, canonicalise Transferable
