+++ types/node/test/globals.ts
@@ -60,17 +60,6 @@ declare var RANDOM_GLOBAL_VARIABLE: true;
     }
 }
 
-// structuredClone
-{
-    structuredClone(123); // $ExpectType 123
-    structuredClone("hello"); // $ExpectType "hello"
-    structuredClone({ test: 123 }); // $ExpectType { test: number; }
-    structuredClone([{ test: 123 }]); // $ExpectType { test: number; }[]
-
-    const arrayBuffer = new ArrayBuffer(0);
-    structuredClone({ test: arrayBuffer }, { transfer: [arrayBuffer] }); // $ExpectType { test: ArrayBuffer; }
-}
-
 // Array.prototype.at()
 {
     const mutableArray = ["a"];
